Updated after merge with validFunctionParameters change
